+++ CS_143/src/EventGenerators/Host.h
@@ -44,8 +44,8 @@ public:
     void respondTo(PacketEvent new_event);
     void respondTo(FlowEvent flow_event);
     void respondTo(UnackEvent unack_event);
-    void sendAndQueueResend(std::shared_ptr<Packet> pkt, float time, float delay);
-    void send(std::shared_ptr<Packet> pkt, float time);
+    void sendAndQueueResend(std::shared_ptr<Packet> pkt, double time, float delay);
+    void send(std::shared_ptr<Packet> pkt, double time);
 
     void respondToSynUnackEvent(UnackEvent unack_event);
     void respondToFinUnackEvent(UnackEvent unack_event);
